refactor: apply design patterns to eliminate config duplication
- Strategy: extract shared configurators into config/common/ so
  cluster and namespaced scopes delegate to a single source of truth
- Template Method: unify GetProvider/GetProviderNamespaced via
  buildProvider(rootGroup) in config/provider.go
- Builder: introduce optionsBuilder in cmd/provider/main.go to
  construct tjcontroller.Options once, varying only the Provider field
- Factory Method: add config/conndetails/ package with StringKeys and
  IndexedSlice factories for AdditionalConnectionDetailsFn
- Add Codecov coverage badge to README

Net result: -942 lines duplicated, +137 lines shared. All existing
tests pass with coverage maintained at 100% on business logic.
